% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendataloader-pdf.R
\name{opendataloader_pdf}
\alias{opendataloader_pdf}
\title{Import the opendataloader_pdf Python module}
\usage{
opendataloader_pdf(delay_load = FALSE)
}
\arguments{
\item{delay_load}{Whether reticulate should delay loading Python until the
module is first used.}
}
\value{
The imported Python module.
}
\description{
Import the opendataloader_pdf Python module
}
